requirements.txt
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-huggingface>=0.0.1
google-genai>=0.1.0
pydantic>=2.0.0
langchain-google-genai>=1.0.0
langgraph>=0.2.0